+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/coneKotlinType.kt
@@ -1,5 +1,8 @@
 package com.javiersc.kotlin.compiler.extensions.fir
 
+import com.javiersc.kotlin.compiler.extensions.common.classId
+import org.jetbrains.kotlin.fir.FirSession
+import org.jetbrains.kotlin.fir.plugin.createConeType
 import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.ConeClassLikeLookupTagImpl
 import org.jetbrains.kotlin.fir.symbols.impl.FirCallableSymbol
@@ -9,6 +12,15 @@ import org.jetbrains.kotlin.fir.types.ConeTypeProjection
 import org.jetbrains.kotlin.fir.types.coneType
 import org.jetbrains.kotlin.fir.types.impl.ConeClassLikeTypeImpl
 import org.jetbrains.kotlin.name.ClassId
+
+public inline fun <reified T> FirSession.coneKotlinType(
+    vararg typeArguments: ConeTypeProjection
+): ConeKotlinType = this.coneKotlinType<T>(typeArguments.toList().toTypedArray())
+
+@JvmName("coneKotlinTypeWithTypeArgumentsArray")
+public inline fun <reified T> FirSession.coneKotlinType(
+    typeArguments: Array<ConeTypeProjection> = emptyArray()
+): ConeKotlinType = classId<T>().createConeType(this, typeArguments)
 
 public val FirBasedSymbol<*>.coneKotlinType: ConeKotlinType?
     get() =